# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Task API",
        "description": "API para gestión de proyectos y tareas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Obtener lista de tareas",
                "operationId": "6a93d54730cde509d735e538ab6800f6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filtrar por prioridad",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project_id",
                        "in": "query",
                        "description": "Filtrar por proyecto",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Mi Tarea"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descripción de la tarea"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "priority": {
                                                        "type": "string",
                                                        "example": "high"
                                                    },
                                                    "project_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "due_date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-12-31"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tareas"
                ],
                "summary": "Crear una nueva tarea",
                "operationId": "9b481c0ed5fdfa1a592d70ba39a348ee",
                "requestBody": {
                    "description": "Datos de la tarea a crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "status",
                                    "priority",
                                    "project_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Título de la tarea",
                                        "type": "string",
                                        "example": "Nueva Tarea"
                                    },
                                    "description": {
                                        "description": "Descripción de la tarea",
                                        "type": "string",
                                        "example": "Descripción detallada de la tarea"
                                    },
                                    "status": {
                                        "description": "Estado de la tarea",
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "in_progress",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "pending"
                                    },
                                    "priority": {
                                        "description": "Prioridad de la tarea",
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "urgent"
                                        ],
                                        "example": "medium"
                                    },
                                    "project_id": {
                                        "description": "ID del proyecto al que pertenece la tarea",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "due_date": {
                                        "description": "Fecha límite (opcional)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-31"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarea creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Nueva Tarea"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Descripción detallada de la tarea"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "pending"
                                                },
                                                "priority": {
                                                    "type": "string",
                                                    "example": "medium"
                                                },
                                                "project_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "due_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-31"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
...
    },
...
}